+++ src/biodwh2-core/src/test/java/de/unibi/agbi/biodwh2/core/collections/ReverseLongTrieTest.java
@@ -110,6 +110,21 @@ class ReverseLongTrieTest {
     }
 
     @Test
+    void concurrentModificationTest() {
+        final ReverseLongTrie trie = new ReverseLongTrie();
+        trie.add(5L);
+        trie.add(55L);
+        trie.add(555L);
+        trie.add(554L);
+        trie.add(10L);
+        trie.add(3490L);
+        assertDoesNotThrow(() -> {
+            for (final Long value : trie)
+                trie.remove(value);
+        });
+    }
+
+    @Test
     void serializableTest() throws IOException, ClassNotFoundException {
         final ReverseLongTrie trie = new ReverseLongTrie();
         trie.add(321L);